Change the font size on the end game panel and write right time into text file

+++ src/nz/ac/aut/ense701/gameModel/Game.java
@@ -956,7 +956,7 @@ public class Game
         }
 
 
-    public int min,sec;
+    public double min,sec;
     private MusicPlayer mplayer;
 	private Island island;
     private Player player;
